Still working on paging
Just need to quicksave to resume this on my other PC

+++ kernelsrc/include/arch/x86/paging.h
@@ -4,6 +4,6 @@
 
 uint16_t set_pte_flags(bool p, bool rw, bool us, bool pwt, bool pcd, bool a, bool d, bool pat, bool g);
 
-uint16_t set_pte(uint32_t addr, uint16_t flags);
+uint32_t set_pte(uint32_t addr, uint16_t flags);
 
 #endif
